Commit the first step towards playing videos from Python.
cocoavlc.py is taken from the VLC Python bindings examples [0]. It
provides a simple player built with Python. loop.py is the main script
that starts playing a video. I have separated out start_app.py as the
player loses control when the video is started, and I can stop it when
it runs in a separate process.

[0] http://git.videolan.org/?p=vlc/bindings/python.git;a=blob;f=examples/cocoavlc.py;h=29d829920b4208efdd51373bdef243cd6fca9aa0;hb=HEAD