Start the next movie before terminating the current one.
Before the transition between movies was quite jarring as it would stop
one and go back to the desktop before starting the next one. Now we
start the next movie over the current one before stopping the one in the
background. It is not perfect but looks much better already.